configure: always build with -fPIC on Linux
To allow the shared library to be built we must build with -fPIC.
Normally libtool would do this for us provided --disable-shared
is not provided, but as we have to link against the Rust code
which is not a libtool library, we have to go out of our way
to avoid libtool.

This still requires that Suricata be built with --enable-shared
(default). Using --disable-shared will still not allow the shared
library to be built.