detect/content: Consider distance in validation
Ticket: 2982

This commit validates that the content usage in a rule will not exceed
the dsize value.

Values of distance that cause the right edge to be exceeded are
considered an error and the signature will be rejected.

(cherry picked from commit 05e7987f242b3c86573ffc7380bc9e44a2d4e0d6)