detect/content: Consider distance in validation
Ticket: 2982

This commit validates that the content usage in a rule will not exceed
the dsize value.

Values of distance that cause the right edge to be exceeded are
considered an error and the signature will be rejected.

(cherry picked from commit 7eb5fb1826233d7a25720e85baf4ab1baab52c1d)