[AWN-57453] Fix flow hash-table corruption
PROBLEM: Crash when running ixia traffic mixes

CAUSE: TCP 5-tuple re-use caused Suricata to forcefully close off
a flow, and start a new one (not unlike what we do to support
midstream flow logs).  The new flow is added to the head of the
hash bucket (but the previous flow pointer still points to
the flow prior to the old flow).
If we then decided to do a midstream flow log (because the
flow hasn't been initialized with the packet timestamp) then
we'd unlink the flow from the wrong previous flow and
corrupt that hash-table bucket (we'd also add another new
flow to the head of the bucket, which would then end up
referencing the old flow as its next pointer, even though
the old flow is about to be freed).

SOLUTION: Don't even check for flow periodic logs in this
case as we *know* the flow is actually new.