app-layer/pd: only consider actual available data
For size limit checks consider only available data in the whole stream.

The old check would consider too much data if there were temporary gaps,
like when a data packet was in-window but (far) ahead of the expected
segment : the old check considered the gap as if the data was not
missing.