introduce support of DPDK as capture interface in workers mode
Register a new runmode - DPDK. This enables a new flag on Suricata start
(--dpdk).
With the flag given, DPDK runmode is enabled.
Runmode loads the configuration and then initializes EAL.
If successful, it configures the NICs according to the configuration file.
After that, worker threads are initialized and then are in continuous receive
loop.