netmap: Netmap API 14 support
This commit provides support for Netmap API version 14+.

We thank @bmeeks8 for his contributions -- much of the content in this
commit is his.

This commit is in *DRAFT* mode and should not be used in production
systems.

The most notable change is that multiple threads are created it threads
is configured as "auto".

Multiple thread support is working for
- IPS with Host method, e.g., interface-name and interface-name^
- IDS mode

Multiple thread support is not yet working for other scenarios.