rules: fix flow drop for ip-only sig with noalert
A simple drop rules like drop tcp any any -> any any (noalert; sid:1; rev:1;)
fails to drop packets beyond the first in each direction of the flow.

The cause is that the rule is just evaluated once per flow direction, by the
"IP-only" detection path. In this case the 'noalert' presence causes the drop
not to be applied to the flow.

Bug: #4663