app-layer: disable by default if not in configuration
DNP3, ENIP and HTTP2 are supposed to be disabled
by default. That means the default configuration does it,
but that also means that, if they are not in suricata.yaml,
the protocol should stay disabled.

(cherry picked from commit ea4a509a54e8292a41a496fc9d2b97096de91f77)