rust/app-layer: use generics for ffi iterator access
By creating a trait for state that requires the state
implement a transaciton slice accessor method, we can then
use a default trait method plus a generic function to implement
the ffi code for C to get the trait iterator, removing that
duplicated code from each app-layer implementation.

This process could also be used for the couple of methods we
create using macros.