af-packet: fix tpacket-v2 ref cnt handling
Socket ref cnt is decremented from the v2 release callback. If the
callback would get to ref cnt 0, the packet would not be released
in the kernel, but it would (possibly) close the socket if the
iface was down.

This patch changes the logic to first release the packet to the
kernel and then decrement the ref cnt.