af-packet: remove autofp support
Autofp is a lot slower than workers, with much higher latency when used in IPS
mode. All supported kernels support proper fanout, so there is no reason to
support the added complexity of having multiple threads dealing with a single
ring.

Remove runmode as well as the various synchronization facilities for autofp.

Ticket: #4806.