WIP app/records: initial support
The idea of stream records is that the applayer parsers can tag PDUs and
other arbitrary records in the stream while parsing. These records can then
be inspected from the rule language. This will allow rules that are more
precise and less costly.

The records are stored per direction in the `AppLayerParserState` and will only
be initialized when actual records are in use. The per direction storage has a
fixed size static portion and dynamic support for a larger number. This is done
for effeciency.

When the Stream Buffer slides, records are updated as they use offsets relative
to the stream. A negative offset is used for records that started before the
current window.

Records have events to inspect/log parser errors that don't fit the TX model.