WIP detect: implement 'record' keyword
Implement a special sticky buffer to select records for inspection.

This keyword takes an argument to specify the per protocol record type:

    record:<app proto name>.<specific record name>

Examples:

    tls.pdu
    smb.smb2.hdr
    smb.smb3.data

Consider a rule like:

    alert tcp ... flow:to_server; content:"|ff|SMB"; content:"some smb 1 issue";

this will scan all toserver TCP traffic, where it will only be limited by a port,
depending on how rules are grouped.

With this work we'll be able to do:

    alert smb ... flow:to_server; record:smb.smb1.data; content:"some smb 1 issue";

This rule will only inspect the data portion of SMB1 records. It will not affect
any other protocol, and it won't need special patterns to "search" for the
SMB1 record in the raw stream.